+++ lib/src/actions/entity-action.spec.ts
@@ -48,6 +48,12 @@ describe('EntityActionFactory', () => {
     expect(action.type).toBe(expectedFormat);
   });
 
+  it('should format type with given label instead of the entity name', () => {
+    const label = 'Hero - Label Test';
+    const action = factory.create('Hero', EntityOp.QUERY_ALL, null, label);
+    expect(action.type).toContain(label);
+  });
+
   it('can re-format generated action.type with custom #formatActionType()', () => {
     factory.formatActionType = (op, entityName) =>
       `${entityName}_${op}`.toUpperCase();